+++ src/components/navigation/navigation.tsx
@@ -1,25 +1,27 @@
-import * as React from "react";
-import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu";
-import * as AccordionPrimitive from "@radix-ui/react-accordion";
-import { cva, type VariantProps } from "class-variance-authority";
-import { ChevronDownIcon } from "@stash-ui/regular-icons";
+import * as React from 'react';
+import * as NavigationMenuPrimitive from '@radix-ui/react-navigation-menu';
+import * as AccordionPrimitive from '@radix-ui/react-accordion';
+import { cva, type VariantProps } from 'class-variance-authority';
+import { ChevronDownIcon } from '@stash-ui/regular-icons';
 
-import { cn } from "@/lib/utils";
+import { cn } from '@/lib/utils';
 
 const navigationVariants = cva(
-  "h-[40px] inline-flex  items-center justify-between w-full p-3 rounded-lg cursor-pointer text-sm font-normal",
+  'h-[40px] inline-flex  items-center justify-between hover:no-underline w-full p-[12px] rounded-lg cursor-pointer text-sm font-normal',
   {
     variants: {
       variant: {
         highlighted:
-          "shadow shadow-[0px_0px_0px_3px_rgba(14,159,110,0.12)] hover:bg-list-highlighted text-success-foreground",
+          'shadow shadow-[0px_0px_0px_3px_rgba(14,159,110,0.12)] hover:bg-list-highlighted hover:text-success-foreground text-success-foreground',
         default:
-          "text-tertiary-foreground hover:text-active-foreground hover:bg-list-actived",
-        disabled: "text-tertiary-foreground opacity-[0.45] cursor-not-allowed",
+          'text-tertiary-foreground hover:text-active-foreground hover:bg-list-actived',
+        disabled: 'text-tertiary-foreground opacity-[0.45] cursor-not-allowed',
+        active:
+          'bg-active-menu text-active-foreground  font-semibold hover:text-active-foreground',
       },
     },
     defaultVariants: {
-      variant: "default",
+      variant: 'default',
     },
   }
 );
@@ -37,18 +39,19 @@ function NavigationLink({
   ...props
 }: NavigationLinkProps) {
   return (
-    <NavigationMenuPrimitive.Root>
+    <NavigationMenuPrimitive.Root className={cn('w-full', className)}>
       <NavigationMenuPrimitive.Link
         data-testid='navigation-link'
         className={cn(
           navigationVariants({ variant }),
           className,
+          'justify-start hover:no-underline',
           isActive &&
             !variant &&
-            "text-active-foreground bg-active-menu font-semibold",
+            'text-active-foreground bg-active-menu font-semibold',
           isActive &&
-            variant === "highlighted" &&
-            "text-success-foreground bg-list-highlighted font-semibold"
+            variant === 'highlighted' &&
+            'text-success-foreground bg-list-highlighted font-semibold'
         )}
         {...props}
       />
@@ -74,13 +77,12 @@ function NavigationItem({
         data-testid='navigation-item'
         className={cn(
           navigationVariants({ variant }),
-          className,
           isActive &&
             !variant &&
-            "text-active-foreground bg-active-menu font-semibold",
+            'text-active-foreground bg-active-menu font-semibold',
           isActive &&
-            variant === "highlighted" &&
-            "text-success-foreground bg-list-highlighted font-semibold"
+            variant === 'highlighted' &&
+            'text-success-foreground bg-list-highlighted font-semibold'
         )}
         {...props}
       />
@@ -90,16 +92,18 @@ function NavigationItem({
 
 export interface NavigationSubItemProps
   extends React.ComponentProps<typeof NavigationMenuPrimitive.Link> {
-  position?: "first" | "middle" | "last" | "only";
+  position?: 'first' | 'middle' | 'last' | 'only';
+  activeItem?: boolean;
 }
 
 function NavigationSubItem({
   className,
   position,
+  activeItem,
   ...props
 }: NavigationSubItemProps) {
   const renderLeftIconPosition = (position: string) => {
-    if (position === "first") {
+    if (position === 'first') {
       return (
         <svg
           width='40'
@@ -122,7 +126,7 @@ function NavigationSubItem({
         </svg>
       );
     }
-    if (position === "last") {
+    if (position === 'last') {
       return (
         <svg
           width='40'
@@ -145,7 +149,7 @@ function NavigationSubItem({
         </svg>
       );
     }
-    if (position === "only") {
+    if (position === 'only') {
       return (
         <svg
           width='40'
@@ -197,14 +201,15 @@ function NavigationSubItem({
       <NavigationMenuPrimitive.Link
         data-testid='navigation-sub-item'
         className={cn(
-          "h-[40px] flex items-center flex-start w-full px-3 cursor-pointer text-xs font-normal text-primary hover:bg-list-hover transition duration-300 ease-in-out",
+          'h-[40px] flex items-center hover:no-underline flex-start w-full px-3 cursor-pointer text-xs font-normal text-primary hover:bg-list-hover transition duration-300 ease-in-out',
           className
         )}
         {...props}
       >
         <div
           className={cn(
-            "w-full opacity-60 hover:opacity-100 flex items-center text-tertiary-foreground flex-start transition-all duration-300 z-9"
+            'w-full opacity-60 hover:opacity-100 flex items-center text-tertiary-foreground flex-start transition-all duration-300 z-9',
+            activeItem ? 'font-semibold' : 'font-normal'
           )}
         >
           {position && renderLeftIconPosition(position)}
@@ -216,39 +221,46 @@ function NavigationSubItem({
 }
 
 interface NavigationTriggerProps
-  extends Omit<AccordionPrimitive.AccordionItemProps, "value"> {
-  items: { label: string; href: string }[];
+  extends Omit<AccordionPrimitive.AccordionItemProps, 'value'> {
+  items: { label: string; href: string; value: string | number }[];
+  onClickItem: (item: { label: string; href: string }) => void;
+  activeItem?: string | number;
 }
 
 function NavigationTrigger({
   className,
   items,
   children,
+  onClickItem,
+  activeItem,
   ...props
 }: NavigationTriggerProps) {
   const handlePosition = (itemIndex: number, itemsLength: number) => {
-    if (itemsLength === 1) return "only";
-    else if (itemIndex === 0 && itemsLength > 1) return "first";
-    else if (itemIndex === itemsLength - 1) return "last";
-    else return "middle";
+    if (itemsLength === 1) return 'only';
+    else if (itemIndex === 0 && itemsLength > 1) return 'first';
+    else if (itemIndex === itemsLength - 1) return 'last';
+    else return 'middle';
   };
 
+  const handleClickItem = (
+    item: { label: string; href: string; value: string | number },
+    e: React.MouseEvent<HTMLAnchorElement, MouseEvent>
+  ) => {
+    e.stopPropagation();
+    e.preventDefault();
+    onClickItem(item);
+  };
   return (
     <AccordionPrimitive.Root
       type='single'
       collapsible
-      className={className}
       data-testid='navigation-trigger'
+      className='w-full'
     >
-      <AccordionPrimitive.Item
-        {...props}
-        className='AccordionItem'
-        value='accordion'
-      >
+      <AccordionPrimitive.Item {...props} value='accordion'>
         <AccordionPrimitive.Trigger
           className={cn(
-            "h-[40px] inline-flex items-center justify-between w-full cursor-pointer text-sm font-normal text-tertiary-foreground  p-3 rounded-lg  hover:text-active-foreground [&[data-state=open]>svg]:rotate-180 [&[data-state=open]]:bg-active-menu [&[data-state=open]]:text-active-foreground [&[data-state=open]]:font-semibold hover:text-opacity-100 hover:bg-list-actived transition duration-400 ease-in-out",
-            className
+            'w-full h-[40px] inline-flex items-center justify-between w-full cursor-pointer text-sm font-normal text-tertiary-foreground  p-[12px] rounded-lg  hover:text-active-foreground [&[data-state=open]>svg]:rotate-180 [&[data-state=open]]:bg-active-menu [&[data-state=open]]:text-active-foreground [&[data-state=open]]:font-semibold hover:text-opacity-100 hover:bg-list-actived transition duration-400 ease-in-out'
           )}
         >
           {children}
@@ -256,14 +268,17 @@ function NavigationTrigger({
         </AccordionPrimitive.Trigger>
         <AccordionPrimitive.Content
           className={cn(
-            "w-full overflow-hidden data-[state=open]:animate-slide-up data-[state=closed]:animate-slide-down"
+            'w-full overflow-hidden data-[state=open]:animate-slide-up data-[state=closed]:animate-slide-down'
           )}
         >
           {items.map((item, index) => (
             <NavigationSubItem
               href={item.href}
+              onClick={(e) => handleClickItem(item, e)}
               key={index}
+              accessKey={index.toString()}
               position={handlePosition(index, items.length)}
+              activeItem={activeItem === item.value}
             >
               {item.label}
             </NavigationSubItem>
@@ -279,10 +294,10 @@ export interface NavigationListProps
 
 function NavigationList({ ...props }: NavigationListProps) {
   return (
-    <NavigationMenuPrimitive.Root>
+    <NavigationMenuPrimitive.Root className={cn('w-full')}>
       <NavigationMenuPrimitive.List
         className={cn(
-          "group flex list-none flex-col gap-1 w-full overflow-hidden"
+          'group flex list-none flex-col gap-1 w-full overflow-hidden gap-0'
         )}
         {...props}
       />